+++ lint-test.sh
@@ -9,3 +9,10 @@ cargo fmt -- --check
 
 cargo clippy --all-features --all-targets --examples --tests
 cargo test --all-features
+
+for ex in examples/*; do
+    base=$(basename $ex)
+    exname=$(echo $base | cut -d . -f 1)
+    cargo run --example $exname
+done
+